Prevent summary generator from creating invalid HTML
Problem: given a source document `example.md`:

    This is a summary.

    READMORE

    The extended article continues here...

Middleman first renders the markdown, then splits on the `READMORE`
token to create a summary. The rendered document would appear as:

    <p>This is a summary.</p>

    <p>READMORE</p>

    <p>The extended article continues here...</p>

Splitting the rendered text at the `READMORE` token:
https://github.com/middleman/middleman-blog/blob/69e214d58080cc06deb72f29fe6067806870cc7a/lib/middleman-blog/blog_article.rb#L93
Produces a summary with an unclosed p tag:

    '<p>This is a summary.</p>\n<p>'

My temporary fix is to use `<p>READMORE</p>` as the summary_separator.
It seems to work for Markdown files, but may break with other files.